Create C++ Custom template
you can use the code on line 28 to print the elements of the array along with the index so it's easier to debug and understand your code. It works like this : If the Array name is a and the index starts from 0 to n - 1, it will print the contents of the array like this -> a[0] = ____ (if i = 0).  You can change the data types in the #define if you need a longer range.